# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Mercado Agrícola — API de Simulación
  version: 1.0.0
  summary: API REST del servidor autoritativo de un mercado agrícola simulado.
  description: |
    Servidor autoritativo de estado para una simulación de mercado de productos
    agrícolas con ~100 agentes concurrentes (productores primarios,
    transformadores, consumidores y traders). Esta API es consumida indistintamente
    por agentes automatizados (reglas o ML) y por clientes humanos: ambos usan
    el mismo contrato.

    ## Modelo de transporte

    - **REST sobre HTTPS** para todas las operaciones síncronas: registro,
      autenticación, consultas de estado, colocación/cancelación de órdenes
      e inicio/cancelación de procesos de transformación.
    - **WebSocket** (`/v1/ws`) para notificaciones push del servidor al agente
      (`order_executed`, `order_expired`, `transformation_completed`,
      `bankruptcy_notice`, `agent_bankrupt`, `trade_printed`). El canal WS no
      acepta comandos: es unidireccional servidor→cliente más heartbeats.
      Su esquema se documenta como complemento al final (sección
      `x-websocket-channel`); está fuera del modelo formal de OpenAPI 3.1.

    ## Unidades y precisión

    Todas las cantidades del dominio se transmiten como **enteros**:

    - **Cantidades de producto**: enteros en *centésimas* de la unidad declarada
      en el producto (ej. `1500` representa `15.00 kg`).
...
      unidades monetarias).

...

    ## Tiempo

...

    ## Autenticación

...
servers:
...
paths:
...
  /catalog/products/{product_id}:
    get:
      tags: [catalog]
      summary: Detalle de un producto
      security: []
      parameters:
        - $ref: '#/components/parameters/ProductIdPath'
      responses:
        '200':
          description: Producto.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404': { $ref: '#/components/responses/NotFound' }
  /catalog/recipes:
    get:
      tags: [catalog]
      summary: Listar recetas del catálogo
      description: |
        Devuelve todas las recetas con sus insumos. Las recetas con lista de
        insumos vacía modelan producción primaria desde cero.
      security: []
      parameters:
        - in: query
          name: output_product_id
          description: Filtrar recetas que producen un producto específico.
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Lista de recetas.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Recipe' }
  /catalog/recipes/{recipe_id}:
    get:
      tags: [catalog]
      summary: Detalle de una receta
      security: []
      parameters:
        - $ref: '#/components/parameters/RecipeIdPath'
      responses:
        '200':
          description: Receta.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Recipe' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /agents/me/installations:
    get:
      tags: [agent]
      summary: Instalaciones compradas del agente
      description: |
        Devuelve las instalaciones que el agente ha comprado (economía de
        instalaciones, ADR-021), cada una con su nivel (nº de hectáreas / líneas
        de producción = presupuesto de concurrencia COMPARTIDO por las recetas
        del tipo), los procesos del tipo en curso, los huecos libres y el precio
        de la siguiente mejora.
      responses:
        '200':
          description: Lista de instalaciones.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/InstallationStatus' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [agent]
      summary: Comprar o mejorar una instalación
      description: |
        Compra la instalación (si no la posee, la crea en nivel 1) o la sube un
        nivel. El precio de pasar del nivel `k` al `k+1` es
        `floor(base_price_cents × (growth_bps/10000)^k)`. El pago se debita del
        capital del agente y se **acredita al banco central** vía el ledger
        append-only (ADR-021), sin sumidero: la conservación monetaria se mantiene.

        `expected_current_level` (opcional) da concurrencia optimista: si se envía
        y no coincide con el nivel actual, falla con `conflict_state` sin cobrar.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AcquireInstallationRequest' }
      responses:
        '201':
          description: Instalación comprada/mejorada.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcquireInstallationResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: |
            Prohibido. Causas vía `errors[*].code`: `agent_bankrupt`,
            `installation_role_mismatch` (el tipo no corresponde al rol).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: |
            `unknown_installation_type`: el `installation_type` no existe.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: |
            `conflict_state`: `expected_current_level` no coincide con el nivel actual.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: |
            Validación de dominio. Causas vía `errors[*].code`:
            `insufficient_capital` (no alcanza para el precio),
            `installation_max_level` (ya en nivel máximo).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token JWT obtenido vía `/auth/login`, `/auth/register` o
        `/auth/refresh`. Stateless, vida corta (sugerido 15 min). Enviar como
        `Authorization: Bearer <token>`.

  # --------------------------------------------------------------------------
  # PARÁMETROS REUTILIZABLES
  # --------------------------------------------------------------------------
  parameters:
    ProductIdPath:
      in: path
      name: product_id
      required: true
      schema: { type: string, format: uuid }
    RecipeIdPath:
      in: path
      name: recipe_id
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Falta token o el token es inválido/expirado.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    NotFound:
      description: Recurso no encontrado.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: |
        Formato RFC 7807 con extensión `errors` para listar causas múltiples
        cuando aplique (típicamente en validación de dominio).
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI identificadora del tipo de problema.
          example: "https://errors.mercado-agricola/insufficient-capital"
        title:
          type: string
          example: "Capital insuficiente"
        status:
          type: integer
          example: 422
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          items:
            type: object
            required: [code, message]
            properties:
              code:
                type: string
                example: "insufficient_capital"
              field:
                type: string
                nullable: true
              message:
                type: string

    # ---- Auth ------------------------------------------------------------
...
    ProductCategory:
      type: string
      enum: [raw_primary, intermediate, final_consumption]
      description: |
        Categoría informativa del producto. **No restrictiva**: cualquier
        agente puede operar cualquier producto.
    Product:
      type: object
      required: [product_id, key, name, unit, category, created_at]
      properties:
        product_id: { type: string, format: uuid }
        key:
          type: string
          description: |
            Identificador estable del catálogo (ej. `trigo`, `harina`). A
            diferencia de `product_id` (UUID generado en cada seed), `key` es
            constante entre despliegues; los clientes pueden anclar su
            configuración (precios base, universos) a este valor.
        name: { type: string }
        unit:
          type: string
          description: Unidad de medida del producto (ej. `kg`, `L`, `cabezas`).
        category: { $ref: '#/components/schemas/ProductCategory' }
        created_at: { type: string, format: date-time }
    RecipeInput:
      type: object
      required: [product_id, qty_required_cent]
      properties:
        product_id: { type: string, format: uuid }
        qty_required_cent:
          type: integer
          format: int64
          minimum: 1
          description: Cantidad requerida por ejecución, en centésimas de la unidad del producto.
    Recipe:
      type: object
      required:
        - recipe_id
        - name
        - output_product_id
        - output_qty_cent
        - duration_seconds
        - wage_rate_cents_per_sec
        - installation_type_id
        - inputs
        - created_at
      properties:
        recipe_id: { type: string, format: uuid }
        name: { type: string }
        output_product_id: { type: string, format: uuid }
        output_qty_cent:
          type: integer
          format: int64
          minimum: 1
          description: Cantidad producida por ejecución, en centésimas de la unidad del producto resultante.
        duration_seconds:
          type: integer
          format: int64
          minimum: 1
          description: Duración de **una** ejecución en segundos reales (no simulados).
        wage_rate_cents_per_sec:
          type: integer
          format: int64
          minimum: 0
        installation_type_id:
          type: string
          format: uuid
          description: Tipo de instalación requerido para ejecutar la receta (ADR-021).
        inputs:
          type: array
          description: Lista de insumos. Vacía para recetas de producción primaria desde cero.
          items: { $ref: '#/components/schemas/RecipeInput' }
        created_at: { type: string, format: date-time }

    # ---- Agentes ---------------------------------------------------------
...
    InstallationStatus:
      type: object
      description: |
        Instalación comprada por el agente (economía de instalaciones, ADR-021).
      required: [installation_type, name, unit_label, level, running, available_slots, next_upgrade_price_cents]
      properties:
        installation_type:
          type: string
          description: '`key` estable del tipo de instalación (ej. `campo`, `metalurgia`).'
        name: { type: string }
        unit_label:
          type: string
          description: Unidad del nivel (ej. `hectareas`, `lineas_produccion`).
        level:
          type: integer
          minimum: 1
          description: |
            Nivel comprado = presupuesto de concurrencia COMPARTIDO por todas las
            recetas del tipo (nº de hectáreas / líneas de producción).
        running:
          type: integer
          minimum: 0
          description: Procesos `running` del tipo actualmente.
        available_slots:
          type: integer
          minimum: 0
          description: '`level - running`. Calculado por el servidor.'
        next_upgrade_price_cents:
          type: integer
          format: int64
          minimum: 0
          nullable: true
          description: Precio de la siguiente mejora, o `null` si ya está en nivel máximo.
...
    AcquireInstallationRequest:
      type: object
      required: [installation_type]
      properties:
        installation_type:
          type: string
          description: '`key` del tipo a comprar/mejorar.'
        expected_current_level:
          type: integer
          minimum: 0
          description: |
            Opcional (concurrencia optimista): si se envía y no coincide con el
            nivel actual, la operación falla con `conflict_state` sin cobrar.
    AcquireInstallationResponse:
      allOf:
        - $ref: '#/components/schemas/InstallationStatus'
        - type: object
          required: [amount_charged_cents]
          properties:
            amount_charged_cents:
              type: integer
              format: int64
              minimum: 0
              description: Precio cobrado por esta compra/mejora (acreditado al banco).
...
security:
  - bearerAuth: []

# ============================================================================
# CANAL WEBSOCKET (extensión informativa, fuera del modelo REST de OpenAPI)
# ============================================================================
...